Clean and comment
