+++ syft/frameworks/torch/tensors/interpreters/precision.py
@@ -744,6 +744,15 @@ class FixedPrecisionTensor(AbstractTensor):
         Forward the .share() command to the child tensor, and reconstruct a new
         FixedPrecisionTensor since the command is not inplace and should return
         a new chain
+
+        Args:
+            *owners: the owners of the shares of the resulting AdditiveSharingTensor
+            field: the field size in which the share values live
+            crypto_provider: the worker used to provide the crypto primitives used
+                to perform some computations on AdditiveSharingTensors
+
+        Returns:
+            A FixedPrecisionTensor whose child has been shared
         """
         if field is None:
             field = self.field
@@ -760,6 +769,10 @@ class FixedPrecisionTensor(AbstractTensor):
         return tensor
 
     def share_(self, *args, **kwargs):
+        """
+        Performs an inplace call to share. The FixedPrecisionTensor returned is therefore the same,
+        contrary to the classic share version version
+        """
         self.child = self.child.share_(*args, no_wrap=True, **kwargs)
         return self
