+++ test/message/test_plan.py
@@ -10,7 +10,7 @@ import syft as sy
 from syft.generic.pointers.pointer_tensor import PointerTensor
 from syft.generic.frameworks.types import FrameworkTensor
 from syft.messaging.plan import Plan
-from syft.messaging.plan import Procedure
+from syft.messaging.plan.procedure import Procedure
 from syft.serde.serde import deserialize
 from syft.serde.serde import serialize
